+++ tailwind.config.js
@@ -9,4 +9,5 @@ module.exports = {
     extend: {},
   },
   plugins: [],
+  darkMode: "class",
 }
